+++ static/CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Add the ability to delete articles persistently
 
+### Fixed
+- Moved startup processing to the background
+- Clean deleted events out of cache, on startup
+- Sped up comment rendering under articles
+
 
 ## [5.2.0] - 2025-12-01